Fix bug with race condition on download window
The progress window is shown on a separate thread. In order to ensure that the download starts after the progress window is shown, we create an Action to run after things are ready. It would be better if things ran using async/await, but this will suffice as a "fix" for now. In the future, we really need to clean up all this "run on UI thread" stuff or whatnot...