Rework AppCast reducing to use filtering verbiage
Also tweaks a few other usages here and there. We don't want both a "Reducer" and a "Filter" and I prefer the word "Filter".

The delegate method and the IAppCastFilter are now a little redundant. Not sure what to do about that. But the interface API keeps things similar to other APIs.